+++ r_segs.cpp
@@ -399,9 +399,9 @@ R_StoreWallRange
     
     // calculate rw_distance for scale calculation
     rw_normalangle = curline->angle + ANG90;
-	offsetangle = abs((int)rw_normalangle - rw_angle1);
+    offsetangle = abs(int(rw_normalangle-rw_angle1));
     
-	    if (offsetangle > ANG90)
+    if (offsetangle > ANG90)
 	offsetangle = ANG90;
 
     distangle = ANG90 - offsetangle;
@@ -621,7 +621,7 @@ R_StoreWallRange
 	offsetangle = rw_normalangle-rw_angle1;
 	
 	if (offsetangle > ANG180)
-	    offsetangle = -(int)offsetangle;
+	    offsetangle = -offsetangle;
 
 	if (offsetangle > ANG90)
 	    offsetangle = ANG90;
